+++ src/main/java/app/AlternativeMain.java
@@ -2,7 +2,7 @@ package app;
 
 import javax.swing.*;
 
-public class Main {
+public class AlternativeMain {
     public static void main(String[] args) {
         AppBuilder appBuilder = new AppBuilder();
         JFrame application = appBuilder
