+++ src/main/java/interface_adapter/Blackjack/BlackjackController.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.Blackjack;
 
 import use_case.BlackjackInteractor;
