+++ src/main/java/interface_adapter/Blackjack/BlackjackViewModel.java
@@ -1,5 +1,5 @@
 
-package interface_adapter;
+package interface_adapter.Blackjack;
 
 import entity.*;
