+++ src/main/java/interface_adapter/Mines/MinesController.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.Mines;
 
 import use_case.MinesInteractor;
