+++ src/main/java/interface_adapter/Mines/MinesPresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.Mines;
 
 import entity.MinesGame;
 import use_case.MinesOutputBoundary;
