--- src/main/java/interface_adapter/tempinterface.java
@@ -1,4 +1,0 @@
-package interface_adapter;
-
-public class tempinterface {
-}
